pyproject.toml
[tool.poetry]
name = "ouobot"
version = "1.0.0"
description = "An Python based Discord bot project for the purpose of studying."
authors = ["SeanLin <linxs1010@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0.0"
"discord.py" = {extras = ["voice"], version = "^1.7.2"}
youtube_dl = "^2021.4.7"
Pillow = "^8.2.0"
discord-py-slash-command = "^1.1.2"
python-dotenv = "^0.15.0"
wavelink = "^0.9.9"
mariadb = "^1.0.6"
requests = "^2.25.1"
bitlyshortener = "^0.5.3"
discord-components = "^0.5.2"
